requirements.txt
Flask==2.3.3
slackeventsapi==3.0.1
slack-sdk==3.23.0
google-generativeai==0.8.3